+++ doc/changelog.rst
@@ -12,7 +12,8 @@ Changelog
   - Emit a warning or an error in english if the standard deviation is larger
     than 10% and/or the shortest sample is shorter than 1 ms
   - Emit a warning or an error if the shortest sample took less than 1 ms
-  - New :attr:`perf.RunResult.loops` and :attr:`perf.RunResult.inner_loops`
+  - Move metadata from :class:`perf.RunResult` to :class:`perf.Benchmark`
+  - New :attr:`perf.Benchmark.loops` and :attr:`perf.Benchmark.inner_loops`
     attributes (they were previously stored in metadata).
   - The ``TextRunner.result`` attribute has been removed, replaced with
     a new :attr:`~perf.text_runner.TextRunner.metadata` attribute (it's also